added transform control panel

+++ util/math/vec.go
@@ -13,9 +13,9 @@ type Vec2i struct {
 }
 
 type Vec3f struct {
-	X float64
-	Y float64
-	Z float64
+	X float32
+	Y float32
+	Z float32
 }
 
 type Vec3i struct {
